cat_base/requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1

